# src/angular/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Portfolio Management API
  version: 0.1.0
  description: |
    Contrat de l'API qui alimente le front Angular.

    ## État

    Première tranche. Elle couvre le socle dont dépendent le plus d'écrans — comptes, titres,
    positions, transactions, places et calendriers — et sert surtout à **fixer les conventions**
    (pagination, erreurs, concurrence, représentation des montants et des dates) avant
    d'étendre. Les ressources restantes sont listées en fin de fichier.

    ## Conventions CRUD

    Chaque ressource qui a une existence propre expose les cinq opérations :

    | Verbe               | Attendu                                                        |
    |---------------------|----------------------------------------------------------------|
    | `GET` collection    | page filtrée et triée                                          |
    | `POST` collection   | création — `201` + `Location` + corps créé                     |
    | `GET` élément       | représentation complète + `ETag`                               |
    | `PUT` élément       | remplacement intégral — les champs omis reprennent leur défaut  |
    | `PATCH` élément     | modification partielle en JSON Merge Patch — `null` efface      |
    | `DELETE` élément    | retrait — `204`, ou `409` si la ressource est encore référencée |

    `PUT` et `PATCH` coexistent parce qu'ils ne disent pas la même chose : le premier affirme
...

...

...
      occurrences.

...

...

...

  contact:
...
  license:
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /securities:
    get:
      tags: [Titres]
      summary: Lister les titres du référentiel
      description: |
        `tradable` et `followed` sont **deux faits indépendants**, et c'est leur conjonction qui
        forme les deux listes de l'écran Titres :

        | Liste | Requête |
        |---|---|
        | Titres négociables | `?tradable=true&followed=false` |
        | Titres suivis | `?followed=true` |

        Un titre en veille reste autorisé par le référentiel : ce qui le retient, c'est la veille,
        pas un refus. `?tradable=true` seul rend donc **aussi** les titres suivis — c'est voulu, et
        c'est la raison pour laquelle la liste des négociables porte les deux critères.

        Deux sous-ressources `/securities/tradable` et `/securities/followed` auraient figé cette
        composition dans des URL, et interdit de demander ce que ni l'une ni l'autre ne nomme —
        les titres écartés, par exemple, ou la veille d'une seule place.
      operationId: listSecurities
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Query'
        - name: mic
          in: query
          description: Restreint à une place de négociation (code MIC).
          schema: { $ref: '#/components/schemas/Mic' }
        - name: tradable
          in: query
          description: |
            Restreint aux titres que le référentiel autorise à la négociation, ou — à `false` — à
            ceux qu'il en écarte. Absent, le critère ne départage pas.
          schema: { type: boolean }
        - name: followed
          in: query
          description: |
            Restreint aux titres tenus en veille, ou — à `false` — à ceux qui ne le sont pas.
            Absent, le critère ne départage pas.
          schema: { type: boolean }
      responses:
        '200':
          description: Page de titres.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Page'
                  - type: object
                    required: [items]
                    properties:
                      items:
                        type: array
                        items: { $ref: '#/components/schemas/Security' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Titres]
      summary: Inscrire un titre au référentiel
      operationId: createSecurity
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SecurityInput' }
      responses:
        '201': { $ref: '#/components/responses/SecurityCreated' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409':
          description: Un titre porte déjà cet ISIN sur cette place.
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /securities/{securityId}:
    parameters:
      - $ref: '#/components/parameters/SecurityId'
    get:
      tags: [Titres]
      summary: Détail d'un titre
      operationId: getSecurity
      responses:
        '200':
          description: Le titre.
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Security' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Titres]
      summary: Remplacer un titre
      operationId: replaceSecurity
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SecurityInput' }
      responses:
        '200': { $ref: '#/components/responses/SecurityUpdated' }
        '404': { $ref: '#/components/responses/NotFound' }
        '412': { $ref: '#/components/responses/PreconditionFailed' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '428': { $ref: '#/components/responses/PreconditionRequired' }
    patch:
      tags: [Titres]
      summary: Modifier un titre
      operationId: patchSecurity
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema: { $ref: '#/components/schemas/SecurityPatch' }
      responses:
        '200': { $ref: '#/components/responses/SecurityUpdated' }
        '404': { $ref: '#/components/responses/NotFound' }
        '412': { $ref: '#/components/responses/PreconditionFailed' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '428': { $ref: '#/components/responses/PreconditionRequired' }
    delete:
      tags: [Titres]
      summary: Retirer un titre du référentiel
      description: |
        Refusé — `409` — si une position ou une transaction le référence encore : le titre
        cesserait d'être résoluble dans des écritures passées. Pour un titre qu'on ne veut plus
        négocier sans rompre l'historique, passer `tradable` à `false`.
      operationId: deleteSecurity
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '204': { description: Titre retiré. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '412': { $ref: '#/components/responses/PreconditionFailed' }
        '428': { $ref: '#/components/responses/PreconditionRequired' }

  # ────────────────────────────── Positions ──────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ─────────────────────────── Paramètres communs ───────────────────────────
  parameters:
    Page:
      name: page
      in: query
      description: Numéro de page, à partir de 1.
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      description: Taille de page.
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Query:
      name: q
      in: query
      description: Recherche plein texte, portée décrite sur chaque opération.
      schema: { type: string }
...
    SecurityId:
      name: securityId
      in: path
      required: true
      schema: { type: string }
      example: SEC-AAPL
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: ETag de la version que le client croit modifier.
      schema: { type: string }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      description: |
        Clé fournie par le client pour qu'un renvoi après coupure réseau ne crée pas un doublon.
        Le serveur rejoue la réponse d'origine si la clé a déjà été vue.
      schema: { type: string, format: uuid }
  headers:
    ETag:
      description: Version de la ressource, à renvoyer en `If-Match` lors d'une modification.
      schema: { type: string }
    Location:
      description: URL de la ressource créée.
      schema: { type: string, format: uri-reference }

  # ──────────────────────── Réponses de création et de mise à jour ────────────────────────
  # Facteur commun des `201`/`200` d'écriture : même en-têtes, même corps que la lecture. Les
  # décrire une fois évite qu'ils divergent d'un verbe à l'autre.
  responses:
...
    SecurityCreated:
      description: Titre inscrit.
      headers:
        Location: { $ref: '#/components/headers/Location' }
        ETag: { $ref: '#/components/headers/ETag' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Security' }
    SecurityUpdated:
      description: Titre mis à jour.
      headers:
        ETag: { $ref: '#/components/headers/ETag' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Security' }
...
    BadRequest:
      description: Requête mal formée.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: 'https://api.portfolio-management.example/problems/invalid-parameter'
            title: Paramètre invalide
            status: 400
            detail: "`from` n'est pas une date au format aaaa-mm-jj."
            errors:
              - pointer: /from
                message: Format de date attendu — aaaa-mm-jj.
    Unauthorized:
      description: Authentification absente ou invalide.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: 'https://api.portfolio-management.example/problems/unauthorized'
            title: Authentification requise
            status: 401
    NotFound:
      description: Ressource inconnue.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: 'https://api.portfolio-management.example/problems/not-found'
            title: Ressource inconnue
            status: 404
            instance: /accounts/BGM-999
    Conflict:
      description: Conflit avec l'état courant de la ressource.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: 'https://api.portfolio-management.example/problems/still-referenced'
            title: Clôture impossible
            status: 409
            detail: Le compte porte encore 12 positions et un solde espèces non nul.
    PreconditionRequired:
      description: |
        L'`If-Match` est absent, vide, ou porte une étiquette que ce serveur n'a pas pu émettre.
        Distinct du `412`, qui répond à une précondition *fournie mais périmée*.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: 'https://api.portfolio-management.example/problems/precondition-required'
            title: Précondition requise
            status: 428
            detail: Cette opération exige un en-tête If-Match portant l'ETag de la version que vous modifiez.
    PreconditionFailed:
      description: L'`If-Match` fourni ne correspond plus à la version courante.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: 'https://api.portfolio-management.example/problems/stale-version'
            title: Version périmée
            status: 412
            detail: La ressource a été modifiée depuis votre lecture. Rechargez avant de réessayer.
    UnprocessableEntity:
      description: Requête bien formée mais refusée par une règle métier.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: 'https://api.portfolio-management.example/problems/business-rule'
            title: Dénouement antérieur à l'exécution
            status: 422
            errors:
              - pointer: /executions/0/settlementDate
                message: La date de dénouement ne peut précéder la date d'exécution.

  # ───────────────────────────────── Schémas ─────────────────────────────────
  schemas:
...
    Problem:
      description: |
        Erreur au format RFC 9457. `errors` porte le détail champ par champ, ce dont les
        formulaires de saisie ont besoin pour placer le message au bon endroit.
      type: object
      required: [type, title, status]
      properties:
        type: { type: string, format: uri, default: 'about:blank' }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri-reference }
        errors:
          type: array
          items:
            type: object
            required: [pointer, message]
            properties:
              pointer:
                type: string
                description: Pointeur JSON vers le champ fautif.
                example: /executions/0/settlementDate
              message: { type: string }
    Page:
      type: object
      required: [page, pageSize, total]
      properties:
        page: { type: integer }
        pageSize: { type: integer }
        total:
          type: integer
          description: Nombre total d'éléments, tous filtres appliqués.
    CurrencyCode:
      type: string
      description: Code ISO 4217.
      pattern: '^[A-Z]{3}$'
      example: EUR
    Mic:
      type: string
      description: Code ISO 10383 de la place. `OTC` pour le gré à gré.
      example: XPAR
...
    Security:
      description: |
        Un titre du référentiel. Trois familles de propriétés cohabitent, et elles n'ont pas le
        même propriétaire :

        - **le noyau**, de `id` à `quoted` — ce que tout consommateur peut attendre ;
        - **la fiche**, de `rating` à `note` — la due diligence attachée au titre, qui s'édite
          avec lui et que l'écran Titres affiche dans son panneau latéral ;
        - **le dérivé**, `held` et `mandates` — issu des positions, jamais écrit. Ces deux
          propriétés ne sont pas là par confort : c'est d'elles que dépend le refus de suppression
          d'un titre encore référencé.
      type: object
      required: [id, ticker, name, mic, currency]
      properties:
        id: { type: string, example: SEC-AAPL }
        ticker: { type: string, example: AAPL }
        name: { type: string, example: Apple Inc. }
        isin:
          type: string
          pattern: '^[A-Z]{2}[A-Z0-9]{9}[0-9]$'
          example: US0378331005
        mic: { $ref: '#/components/schemas/Mic' }
        place:
          type: string
          example: Nasdaq
          readOnly: true
          description: |
            Nom lisible de la place, **dérivé du MIC**. Il n'est jamais reçu en écriture : deux
            clients qui écriraient « Euronext Paris » et « EURONEXT PARIS » produiraient deux
            libellés pour une même place.
        currency: { $ref: '#/components/schemas/CurrencyCode' }
        assetClass: { type: string, example: Actions }
        region: { type: string, example: Amérique du Nord }
        tradable: { type: boolean }
        quoted:
          type: boolean
          description: |
            Faux pour un instrument non coté en continu — fonds fermé, co-investissement — dont
            la valeur est une valeur liquidative publiée, non un cours.
        followed:
          type: boolean
          description: |
            Le titre est tenu en veille plutôt qu'à l'achat. C'est le **périmètre livré par le
            référentiel** : les mises en suivi décidées dans l'application s'empilent par-dessus,
            côté client, et ne remontent pas ici.

        # ── Fiche de référentiel ──────────────────────────────────────────────
        rating: { type: string, example: 'A− (S&P)' }
        cap:
          type: integer
          minimum: 0
          maximum: 100
          description: >-
            Plafond de concentration, **en pourcentage de l'actif net par compte** — `20`, pas
            `0.20`. Zéro signifie « aucun plafond attribué ». Ce n'est pas un rang de taille :
            le champ homonyme d'une composition d'indice, lui, porte bien une capitalisation.
          example: 20
        liquidity: { type: string, example: 'Élevée · 118 M€ de volume moyen' }
        esg: { type: string, example: Article 8 SFDR }
        domicile: { type: string, example: France }
        complexity: { type: string, example: Non complexe }
        reviewed:
          type: string
          format: date
          description: Date de la dernière revue du titre par le comité.
        reviewedBy: { type: string, example: "Comité d'investissement" }
        note: { type: string }

        # ── Dérivé des positions, en lecture seule ────────────────────────────
        held:
          type: integer
          readOnly: true
          description: Quantité détenue, tous comptes confondus.
        mandates:
          type: array
          readOnly: true
          items: { type: string }
          description: Comptes qui portent ce titre, en position ou en historique.
    SecurityInput:
      description: |
        Représentation complète d'un titre en écriture. Ni `id`, ni `place`, ni `held`, ni
        `mandates` : le premier est attribué, le deuxième est dérivé du MIC, les deux derniers
        viennent des positions.

        L'ISIN est facultatif ; **présent, sa clé de contrôle est vérifiée** et un code qui échoue
        à ce contrôle est refusé en `422`. Le motif seul laisse passer une faute de frappe une
        fois sur dix.
      type: object
      required: [ticker, name, mic, currency]
      properties:
        ticker: { type: string, example: AAPL }
        name: { type: string, example: Apple Inc. }
        isin:
          type: string
          pattern: '^[A-Z]{2}[A-Z0-9]{9}[0-9]$'
        mic: { $ref: '#/components/schemas/Mic' }
        currency: { $ref: '#/components/schemas/CurrencyCode' }
        assetClass: { type: string }
        region: { type: string }
        tradable: { type: boolean, default: true }
        quoted: { type: boolean, default: true }
        followed: { type: boolean, default: false }
        rating: { type: string }
        cap: { type: integer, minimum: 0, maximum: 100 }
        liquidity: { type: string }
        esg: { type: string }
        domicile: { type: string }
        complexity: { type: string }
        reviewed: { type: string, format: date }
        reviewedBy: { type: string }
        note: { type: string }
    SecurityPatch:
      description: |
        Modification partielle en JSON Merge Patch. Un champ **absent** n'est pas touché, un champ
        à **`null`** est effacé.

        `ticker`, `mic` et `currency` n'y figurent pas, et ce n'est pas un oubli : changer le
        mnémonique ou la place d'une ligne de référentiel, c'est en désigner une autre. Un patch
        qui les cite est refusé en `422` plutôt qu'ignoré — le taire ferait croire à l'appelant que
        son écriture a porté. Il en va de même d'un champ inconnu.
      type: object
      additionalProperties: false
      properties:
        name: { type: string }
        isin: { type: ['string', 'null'] }
        assetClass: { type: ['string', 'null'] }
        region: { type: ['string', 'null'] }
        tradable: { type: boolean }
        quoted: { type: boolean }
        followed: { type: boolean }
        rating: { type: ['string', 'null'] }
        cap: { type: integer, minimum: 0, maximum: 100 }
        liquidity: { type: ['string', 'null'] }
        esg: { type: ['string', 'null'] }
        domicile: { type: ['string', 'null'] }
        complexity: { type: ['string', 'null'] }
        reviewed: { type: ['string', 'null'], format: date }
        reviewedBy: { type: ['string', 'null'] }
        note: { type: ['string', 'null'] }

    # -- Positions --------------------------------------------------------
...